# PERF: Minor performance improvements to dtypes