A few minor performance enhancements in dtypes modules
1. Use scalar variants of isinf/isnan in `in_dtype_range`. The scalar versions of these functions are much faster than the ufunc (140ns for scalar vs 950ns for ufunc).
2. Avoid creating array in `get_minimum_dtype`. If passed a non-array, use the builtin min/max to avoid the extra memory/cpu overhead of traversing values 3 times.
```
In [66]: %timeit dtypes.get_minimum_dtype(range(266))
35.3 µs ± 651 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [67]: %timeit get_minimum_dtype(range(266))
13.3 µs ± 172 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```
3. Use `np.asarray` where possible. Benchmarking on my system showed that calling `np.asarray` was much faster when called with an array (90ns vs 336ns) and about the same when called with non-arrays (~870us) compared to existing approach.