@groutr 2023-01-23
I think we consider possibly reverting the changes to get_minimum_dtype. The changes made here make the implicit assumption that the dtypes of all the elements of a list (or any other non-array container) are homogeneous.

If `[-3, 0.1, 3]` is passed to get_minimum_dtype, the existing code would have done the right thing and returned a float dtype. However, the new code would return an integer dtype.

Is it permissible to make the assumption that the min/max values reflect the overall type of the array (or at least document that potential trap in the docstring)? One compromise could be to add a keyword argument to `get_mimimum_dtype` for the user to avoid array creation overhead if desired.
```
get_minimum_dtype([-3, 0, 3], check_all=False)  # I know list is only ints, so only use min/max to determine type (int8)
get_minimum_dtype([-3, .01, 3], check_all=True) # return float32
```